+++ REvoDesign/common/MutantTree.py
@@ -1,5 +1,5 @@
 class MutantTree:
-    def __init__(self, mutant_tree):
+    def __init__(self, mutant_tree: dict):
         self.current_branch_id = ''
         self.current_mutant_id = ''
 
@@ -17,7 +17,6 @@ class MutantTree:
         self.empty = bool(len(self.all_mutant_branch_ids) == 0)
 
         if self.empty:
-            print('Empty MutantTree.')
             return
 
         if not self.current_branch_id:
@@ -38,6 +37,13 @@ class MutantTree:
                 tree_str += f"  Mutant: {mutant_id}\n"
                 tree_str += f"    {str(mutant_obj)}\n"
         return tree_str
+    
+    def __copy__(self):
+        return MutantTree(self.mutant_tree.copy())
+    
+    def __deepcopy__(self):
+        import copy
+        return MutantTree(copy.deepcopy(self.mutant_tree))
 
     def get_branch_index(self, branch_id):
         return self.all_mutant_branch_ids.index(branch_id)